Split TopicRegistry into lite and stl
